Update SSL options for OpenSSL v3 (#113)
* Update minimum Ruby version

still an EOL but everything runs successfully

* [cleanup] whitespace and inline blocks when easy

* [fix] Do not default SSL options

`ca_file` and maybe `ca_path` are handled differently under OpenSSL v3 than in v1. It is a safer approach to leave these values out and allow the system to handle default behavior while keeping the configuration available for developers/projects that need it.